// apps/trade-pro/src/pages/CHART/chart_of_account/type.ts
export type formhistory = {
  key: React.Key;
  Id: number;
  AccountCode: number;
  AccountTitle: string;
  Account_Level: number;
  AccountGroup: string;
  BSNoteId: string;
  PLNoteId: string;
  AccountClass: string;
  AccountTypeId: string;
  Status: any;
  ActionId: any;
};
export type child_account = {
  key: React.Key;
  Sr: number;
  Title: string;
  Action: any;
};